simple commandline arguments

+++ src/stream/mod.rs
@@ -2,7 +2,7 @@ use crate::machine::StateMachine;
 
 #[allow(unused_parens)]
 
-pub fn stream(count: usize, dfa: &mut impl StateMachine) {
+pub fn stream(count: u32, dfa: &mut impl StateMachine) {
     dfa.init();
     for i in (0..count) {
         dfa.change();
